Move subtraction to separate modules
